update apis and added a new api for get user information by admin

+++ arcreco/core/views.py
@@ -1,5 +1,5 @@
 from rest_framework import generics
-from rest_framework.viewsets import ModelViewSet
+from rest_framework_simplejwt.views import TokenObtainPairView
 from django.core.mail import EmailMessage
 from .models import UserProfile, UserCompanyLogo
 from rest_framework.response import Response
@@ -9,7 +9,7 @@ from rest_framework.authtoken.views import ObtainAuthToken
 from rest_framework.permissions import IsAuthenticated, IsAdminUser
 import requests
 from .serializers import UserProfileSerializer, UpdateUserProfileSerializer, UpdateUserProfilePictureSerializer, \
-    AddNewUserProfileSerializer, ChangePasswordSerializer, CompanyLogoSerializer
+    AddNewUserProfileSerializer, ChangePasswordSerializer, CompanyLogoSerializer, TokenObtainPairPatchedSerializer
 from . import s3_upload
 from arcreco import settings
 
@@ -27,7 +27,7 @@ class UserCreateAPIView(generics.CreateAPIView):
             serializer.save()
             return Response({
                 'status': 'success',
-                'message': f"Hello {name}, Check your mail for the credentials."
+                'message': f"Hello {name}"
             }, status=status.HTTP_201_CREATED)
         else:
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
@@ -120,7 +120,7 @@ class UserAddAPIView(generics.ListAPIView, generics.CreateAPIView, generics.Upda
             self.send_email(**data)
             return Response({
                 'status': 'success',
-                'message': f"hello {name}"
+                'message': f"hello {name}, please check your mail for login credentials."
             }, status=status.HTTP_201_CREATED)
         else:
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
@@ -136,12 +136,16 @@ class UserAddAPIView(generics.ListAPIView, generics.CreateAPIView, generics.Upda
         return Response({'status': 'success', 'message': 'profile updated'}, status=status.HTTP_200_OK)
 
     def destroy(self, request, *args, **kwargs):
-        try:
-            user = UserProfile.objects.get(id=request.data.get('id'))
-            user.delete()
-            return Response({'status': 'success', 'message': 'user deleted'}, status=status.HTTP_204_NO_CONTENT)
-        except:
-            return Response({'status': 'failed', 'message': 'user not found'}, status=status.HTTP_404_NOT_FOUND)
+        user_id = request.query_params.get('id')
+        if user_id:
+            try:
+                user = UserProfile.objects.get(id=request.query_params.get('id'))
+                user.delete()
+                return Response({'status': 'success', 'message': 'user deleted'}, status=status.HTTP_204_NO_CONTENT)
+            except:
+                return Response({'status': 'failed', 'message': 'user not found'}, status=status.HTTP_404_NOT_FOUND)
+        else:
+            return Response({'id': ['This field is required.']}, status=status.HTTP_400_BAD_REQUEST)
 
 
 class ChangePasswordView(generics.UpdateAPIView):
@@ -195,3 +199,42 @@ class CompanyLogoView(generics.UpdateAPIView):
                 }, status=status.HTTP_202_ACCEPTED)
         else:
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+
+class TokenObtainPairPatchedView(TokenObtainPairView):
+    """
+    Takes a set of user credentials and returns an access and refresh JSON web
+    token pair to prove the authentication of those credentials.
+    """
+    serializer_class = TokenObtainPairPatchedSerializer
+
+    token_obtain_pair = TokenObtainPairView.as_view()
+
+
+class GetUserByAdminView(generics.ListAPIView):
+    """Get single user details by admin on id"""
+    permission_classes = (IsAuthenticated, IsAdminUser,)
+    queryset = UserProfile.objects.all()
+
+    def get(self, request, *args, **kwargs):
+        user_id = request.query_params.get('id')
+        if user_id:
+            try:
+                user = UserProfile.objects.get(id=request.query_params.get('id'))
+                data = {
+                    'f_name': user.name,
+                    'l_name': user.last_name,
+                    'contact': user.contact,
+                    'email': user.email,
+                    'company_name': user.company_name,
+                    'designation': user.designation,
+                    'department': user.department,
+                    'city': user.city,
+                    'state': user.state,
+                    'country': user.country
+                }
+                return Response({'status': 'success', 'data': data})
+            except:
+                return Response({'status': 'failed', 'message': 'user not found'}, status=status.HTTP_404_NOT_FOUND)
+        else:
+            return Response({'id': ['This field is required.']}, status=status.HTTP_400_BAD_REQUEST)
